+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+- `regimes`: New tax regime for the Czech Republic (CZ)
 - `eu-en16931-v2017`: BR-32 validation requiring taxes on document-level discounts.
 
 ### Changed